Move save and load buttons to TopBar
