+++ editor/src/common/components/Typography.tsx
@@ -1,4 +1,6 @@
 import React from 'react';
 import styled from 'styled-components';
 
-export const Title = styled.h1``;
+export const Title = styled.h1`
+  margin: 0 0 2rem;
+`;
